requirements/base.txt
# requirements/base.txt
# Core dependencies for AutoML Builder

# API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# LLM and Agent Framework
langchain==0.1.0
langgraph==0.0.20
openai==1.6.1
tiktoken==0.5.2

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0
redis==5.0.1

# ML Libraries
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
optuna==3.5.0
mlflow==2.9.2
joblib==1.3.2

# Data Processing
openpyxl==3.1.2
xlrd==2.0.1
pyarrow==14.0.1

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
authlib==1.3.0
httpx==0.25.2

# Frontend
streamlit==1.29.0
streamlit-authenticator==0.2.3
streamlit-chat==0.1.1

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
structlog==23.2.0
tenacity==8.2.3

# requirements/dev.txt
# Development dependencies
-r base.txt

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# Code Quality
black==23.12.0
isort==5.13.2
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Development Tools
ipython==8.18.1
jupyter==1.0.0
watchdog==3.0.0

# requirements/prod.txt
# Production dependencies
-r base.txt

# Production servers
gunicorn==21.2.0
gevent==23.9.1

# Monitoring
prometheus-client==0.19.0
sentry-sdk==1.39.1

# Performance
cachetools==5.3.2
msgpack==1.0.7